+++ internal/auth/middleware/auth_middleware.go
@@ -65,7 +65,6 @@ func JWTAuthMiddleware(tokenSvc *tokenService.TokenService, sessionValidator Ses
 
 		ctx.Set(ContextKeyAccountID, claims.AccountID)
 		ctx.Set(ContextKeyClaims, claims)
-		ctx.Set("jwt_token_string", tokenString)
 		ctx.Next()
 	}
 }